Code rearranged, with the right headers
